api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
python-jose
requests